edge/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
psycopg2-binary==2.9.9
azure-iot-device==2.13.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
